fix(flutter): satisfy format and analyzer contracts

+++ lib/features/analytics/domain/habit_metrics.dart
@@ -120,11 +120,11 @@ abstract final class HabitMetricCalculator {
           available.add(date);
         }
       }
-      final scheduled = schedule.target.clamp(0, available.length) as int;
+      final scheduled = schedule.target.clamp(0, available.length);
       final completed = available
           .where(completedDates.contains)
           .length
-          .clamp(0, scheduled) as int;
+          .clamp(0, scheduled);
       weeks.add(
         WeeklyHabitMetric(
           weekStart: weekStart,
